eme: split-up session.ts into two files, one for session creation, the other for event handling
